bug fixes
+app now renders properly on windows 7 and above systems
+added app's icon and mp3, wma file icons
+on new file open, player instantly switches playback to that file
+tray icon added

+++ demp/src/utility/md_time.h
@@ -17,7 +17,7 @@ namespace Time
 		~Timer();
 
 		f32 targetTime;
-		f32 startTime;
+		Uint32 startTime;
 		f32 currentTime;
 
 		b8 started;
